requirements.txt
python-dotenv>=1.0.0
pydantic>=1.10.7
aiohttp>=3.8.4
agent-framework
agent-framework-devui
azure-core>=1.28.0